Use dynamic import for graphics library
The handler used in the AWS lambda is runAll. However, we have no idea
whether the gloval environment (modified by rune_library.js) is
preserved between lambda calls (or if it was there to begin with). This
change causes the rune library to be loaded everytime the handler is
called, ensuring the pre-condition of having the rune libraries. If
needed,  event.name can be used to load up particular libraries if there
are clashes.

Note that this behavior parallels how the libraries are loaded in the
frontend, having to be loaded into the global scope.