Add grader support for runes (#21)
* Add gitignore

* Add support for JS files

- Needed for WebGL code brought in from cadet-frontend
- Changes build directory to avoid overwriting JS files

* Stop ignoring JS files

* Add lodash dependency

* Add dependencies

* Add graphics and matrix library

* Add rune autograder test cases

* Include rune library and fix run to take in Library

* Fix globals in test_runes

* Add build script and ensure build works

* Remove commented-out line

* Use import instead of require

* Add tests for 3D runes

* Fix broken build script

* Ignore grader.zip

* Add dummy visualisation functions

- No parameters causes a slang runtime error
- Not defining the functions will also cause a runtime error

* Use dynamic import for graphics library

The handler used in the AWS lambda is runAll. However, we have no idea
whether the gloval environment (modified by rune_library.js) is
preserved between lambda calls (or if it was there to begin with). This
change causes the rune library to be loaded everytime the handler is
called, ensuring the pre-condition of having the rune libraries. If
needed,  event.name can be used to load up particular libraries if there
are clashes.

Note that this behavior parallels how the libraries are loaded in the
frontend, having to be loaded into the global scope.

* Update README with rune functions